For business card, display background and labels together
When displaying the business card, group the calls to add the
background bitmap and the labels to the group in the same place in
code, allowing them to display at the same time.

This change has removed the black background around the labels, which
appeared to be caused by a bug, since labels don't have a background
color property.