Add, use `show` method to reduce display jank
This also removes 'with open(...) as ...:' because for correct
functionality the OnDiskBitmap needs to retain the open file handle for
as long as it exists.

In particular, this fixes a problem where switching to the 'business card'
view with a background image would redraw twice.

Calling `show()` also acts as activity, so it'll brighten the screen.
At present, this is not configurable.